move routing implementations to respective router object files + fix a bunch of imports
