+++ vllm/model_executor/layers/fused_moe/__init__.py
@@ -83,12 +83,16 @@ if HAS_TRITON:
         BatchedTritonExperts,
     )
     from vllm.model_executor.layers.fused_moe.fused_moe import (
-        GroupedTopk,
         TritonExperts,
         TritonWNA16Experts,
         fused_experts,
-        fused_topk,
         get_config_file_name,
+    )
+    from vllm.model_executor.layers.fused_moe.fused_topk_router import (
+        fused_topk,
+    )
+    from vllm.model_executor.layers.fused_moe.grouped_topk_router import (
+        GroupedTopk,
     )
     from vllm.model_executor.layers.fused_moe.triton_deep_gemm_moe import (
         TritonOrDeepGemmExperts,
